Renaming components to better reflect intention
