--- ui/vendormgmt-react/src/Components/VendorSearch/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import VendorSearchForm from './form';
-
-const VendorSearch = () => {
-  return (
-    <VendorSearchForm />
-  );
-};
-
-export default VendorSearch;
