+++ ui/vendormgmt-react/src/Components/VendorSearch/resultGrid.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ResultGrid = () => {
+  return (
+    <div></div>
+  );
+};
+
+export default ResultGrid;
